# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Serverless CMF with Full Rest API and RBAC(Role Base Control Access) System.",
        "title": "MayCMF",
        "contact": {
            "name": "eneus"
        },
        "license": {},
        "version": "0.1.0"
    },
    "host": "127.0.0.1:8088",
    "basePath": "/",
    "paths": {
...
        "/api/v1/permissions": {
            "get": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Query data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Paging index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Paging Size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Hide permission (0: don't hide 1: hide)",
                        "name": "hidden",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Parent ID",
                        "name": "parentID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search result: {list:List data,pagination:{current:Page index,pageSize:Page size,total:total number}}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Permission"
                            }
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Create data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Create data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.Permission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Permission"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/permissions.tree": {
            "get": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Query permission tree",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "Whether to include action data (1 is)",
                        "name": "includeActions",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Whether to include resource data (1 is)",
                        "name": "includeResources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search result: {list: List data}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.PermissionTree"
                            }
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/permissions/{id}": {
            "get": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Query specified data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Permission"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message:Resource does not exist.}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Update data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.Permission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Permission"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Manage Permission"
                ],
                "summary": "Delete data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error item",
                    "type": "object",
                    "$ref": "#/definitions/schema.HTTPErrorItem"
                }
            }
        },
        "schema.HTTPErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "schema.HTTPStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "status(OK)",
                    "type": "string"
                }
            }
        },
...
        "schema.Permission": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "actions": {
                    "description": "Action list",
                    "type": "object",
                    "$ref": "#/definitions/schema.PermissionActions"
                },
                "created_at": {
                    "description": "Creation time",
                    "type": "string"
                },
                "creator": {
                    "description": "Creator",
                    "type": "string"
                },
                "hidden": {
                    "description": "Hide Permission (0: don't hide 1: hide)",
                    "type": "integer"
                },
                "icon": {
                    "description": "Permission icon",
                    "type": "string"
                },
                "name": {
                    "description": "Permission name",
                    "type": "string"
                },
                "parent_id": {
                    "description": "Parent ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "Parent path",
                    "type": "string"
                },
                "record_id": {
                    "description": "Record ID",
                    "type": "string"
                },
                "resources": {
                    "description": "Resource list",
                    "type": "object",
                    "$ref": "#/definitions/schema.PermissionResources"
                },
                "router": {
                    "description": "Access routing",
                    "type": "string"
                },
                "sequence": {
                    "description": "Sort value",
                    "type": "integer"
                }
            }
        },
...
        "schema.PermissionActions": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "code": {
                        "description": "Action number",
                        "type": "string"
                    },
                    "name": {
                        "description": "Action name",
                        "type": "string"
                    }
                }
            }
        },
...
        "schema.PermissionResources": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "code": {
                        "description": "Resource number",
                        "type": "string"
                    },
                    "method": {
                        "description": "Request method",
                        "type": "string"
                    },
                    "name": {
                        "description": "Resource Name",
                        "type": "string"
                    },
                    "path": {
                        "description": "Request path",
                        "type": "string"
                    }
                }
            }
        },
        "schema.PermissionTree": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "actions": {
                    "description": "Action list",
                    "type": "object",
                    "$ref": "#/definitions/schema.PermissionActions"
                },
                "children": {
                    "description": "Child tree",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PermissionTree"
                    }
                },
                "hidden": {
                    "description": "Hide Permission (0: don't hide 1: hide)",
                    "type": "integer"
                },
                "icon": {
                    "description": "Permission icon",
                    "type": "string"
                },
                "name": {
                    "description": "Permission name",
                    "type": "string"
                },
                "parent_id": {
                    "description": "Parent ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "Parent path",
                    "type": "string"
                },
                "record_id": {
                    "description": "Record ID",
                    "type": "string"
                },
                "resources": {
                    "description": "Resource list",
                    "type": "object",
                    "$ref": "#/definitions/schema.PermissionResources"
                },
                "router": {
                    "description": "Access routing",
                    "type": "string"
                },
                "sequence": {
                    "description": "Sort value",
                    "type": "integer"
                }
            }
        },
...
    }
}